--- a/api/types.go
+++ b/api/types.go
@@ -47,7 +47,6 @@ type DockerContainer struct {
 	DockerComposeFile string
 	ContainerID       string // this assumes that there can only be one container per docker-compose service
 	LogMedium         io.Writer
-	Debug             bool
 }
 
 func (dc *DockerContainer) UpdateContainerID(containerID string) {