- dont output to stdout unless we are in debug mode

// old benchmarks(on master branch):
BenchmarkCreateContainer-32         	  300000	      4855 ns/op
BenchmarkContainerStart-32          	10000000	       187 ns/op
BenchmarkContainerLogs-32           	  200000	      7188 ns/op
BenchmarkPullDockerImage-32         	  200000	      7512 ns/op
BenchmarkBuildDockerImage-32        	  200000	     10075 ns/op
BenchmarkGetNetwork-32              	 5000000	       263 ns/op
BenchmarkConnectNetwork-32          	 3000000	       389 ns/op
BenchmarkCreateDockerVolume-32      	  500000	      3832 ns/op

// new benchmarks(this branch):
BenchmarkCreateContainer-32         	  300000	      4927 ns/op
BenchmarkContainerStart-32          	10000000	       187 ns/op
BenchmarkContainerLogs-32           	  300000	      4579 ns/op
BenchmarkPullDockerImage-32         	  200000	      5928 ns/op
BenchmarkBuildDockerImage-32        	  200000	      9793 ns/op
BenchmarkGetNetwork-32              	 5000000	       258 ns/op
BenchmarkConnectNetwork-32          	 5000000	       385 ns/op
BenchmarkCreateDockerVolume-32      	  500000	      3136 ns/op